editoast: authz: use `Role` enum directly with the `fga` client
Avoids a useless conversion between the former `model::Role` and
`role::Role`.  Also removes a weird parsing step.
